backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic[email]==2.4.2
motor==3.3.1
python-dotenv==1.0.1 
uvloop==0.19.0
httptools==0.6.1
pydantic-settings==2.2.1
rich==13.7.0
bcrypt==4.1.2
python-jose==3.3.0
passlib==1.7.4
